+++ test/ds/test_local.py
@@ -249,6 +249,7 @@ class LocalFilePatternSourceTest(unittest.TestCase):
             new_ds = data_source.make_local('from_local_to_local', None,
                                             (datetime.datetime(1978, 11, 14, 0, 0),
                                              datetime.datetime(1978, 11, 15, 23, 59)))
+            self.assertIsNotNone(new_ds)
             self.assertEqual(new_ds.id, 'local.from_local_to_local')
             self.assertEqual(new_ds.temporal_coverage(), TimeRangeLike.convert(
                 (datetime.datetime(1978, 11, 14, 0, 0),
@@ -269,6 +270,7 @@ class LocalFilePatternSourceTest(unittest.TestCase):
                                                            (datetime.datetime(1978, 11, 14, 0, 0),
                                                             datetime.datetime(1978, 11, 15, 23, 59)),
                                                            None, ['sm'])
+            self.assertIsNotNone(new_ds_w_one_variable)
             self.assertEqual(new_ds_w_one_variable.id, 'local.from_local_to_local_var')
             data_set = new_ds_w_one_variable.open_dataset()
             self.assertSetEqual(set(data_set.variables), {'sm', 'lat', 'lon', 'time'})
@@ -277,7 +279,37 @@ class LocalFilePatternSourceTest(unittest.TestCase):
                                                      (datetime.datetime(1978, 11, 14, 0, 0),
                                                       datetime.datetime(1978, 11, 15, 23, 59)),
                                                      "10,10,20,20", ['sm'])  # type: LocalDataSource
+            self.assertIsNotNone(new_ds_w_region)
             self.assertEqual(new_ds_w_region.id, 'local.from_local_to_local_region')
             self.assertEqual(new_ds_w_region.spatial_coverage(), PolygonLike.convert("10,10,20,20"))
             data_set = new_ds_w_region.open_dataset()
             self.assertSetEqual(set(data_set.variables), {'sm', 'lat', 'lon', 'time'})
+
+            no_data = data_source.make_local('no_data', None,
+                                             (datetime.datetime(2020, 11, 14, 0, 0),
+                                              datetime.datetime(2020, 11, 15, 23, 59)))
+            self.assertIsNone(no_data)
+
+    def test_remove_data_source_by_id(self):
+
+            data_sources = self._local_data_store.query('local_w_temporal')
+            data_sources_len_before_remove = len(data_sources)
+            self.assertGreater(data_sources_len_before_remove, 0)
+
+            with unittest.mock.patch.object(os, 'remove', return_value=None):
+                self._local_data_store.remove_data_source('local_w_temporal')
+            data_sources = self._local_data_store.query('local_w_temporal')
+            data_sources_len_after_remove = len(data_sources)
+            self.assertGreater(data_sources_len_before_remove, data_sources_len_after_remove)
+
+    def test_remove_data_source(self):
+
+            data_sources = self._local_data_store.query('local_w_temporal')
+            data_sources_len_before_remove = len(data_sources)
+            self.assertGreater(data_sources_len_before_remove, 0)
+
+            with unittest.mock.patch.object(os, 'remove', return_value=None):
+                self._local_data_store.remove_data_source(data_sources[0])
+            data_sources = self._local_data_store.query('local_w_temporal')
+            data_sources_len_after_remove = len(data_sources)
+            self.assertGreater(data_sources_len_before_remove, data_sources_len_after_remove)